# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CasinoKing Blog API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Article {
  id: string;
  title: string;
  slug: string;
  excerpt?: string | null;
  content?: string | null;
  category?: string | null;
  image?: string | null;
  author?: string | null;
  authorBio?: string | null;
  authorAvatar?: string | null;
  focusKeyword?: string | null;
  metaDescription?: string | null;
  date?: string | null;
  readTime?: string | null;
  featured?: boolean | null;
  published?: boolean | null;
  createdAt?: string | null;
}

export interface ArticleListResponse {
  articles: Article[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export interface Affiliate {
  id: string;
  name: string;
  logo?: string | null;
  rating?: number | null;
  bonus?: string | null;
  description?: string | null;
  link?: string | null;
  badge?: string | null;
  position?: number | null;
  active?: boolean | null;
}

export interface CategoryCount {
  category: string;
  count: number;
}

export interface ErrorResponse {
  error: string;
}

export type GetArticlesParams = {
  page?: number;
  limit?: number;
  category?: string;
  featured?: boolean;
};

export type GetAllSlugs200 = {
  slugs: string[];
};

export type GetCategories200 = {
  categories: CategoryCount[];
};

export type GetAffiliates200 = {
  affiliates: Affiliate[];
};