# lib/api-zod/src/generated/types/affiliate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CasinoKing Blog API
 * OpenAPI spec version: 0.1.0
 */

export interface Affiliate {
  id: string;
  name: string;
  logo?: string | null;
  rating?: number | null;
  bonus?: string | null;
  description?: string | null;
  link?: string | null;
  badge?: string | null;
  position?: number | null;
  active?: boolean | null;
}