# lib/api-zod/src/generated/types/article.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CasinoKing Blog API
 * OpenAPI spec version: 0.1.0
 */

export interface Article {
  id: string;
  title: string;
  slug: string;
  excerpt?: string | null;
  content?: string | null;
  category?: string | null;
  image?: string | null;
  author?: string | null;
  authorBio?: string | null;
  authorAvatar?: string | null;
  focusKeyword?: string | null;
  metaDescription?: string | null;
  date?: string | null;
  readTime?: string | null;
  featured?: boolean | null;
  published?: boolean | null;
  createdAt?: string | null;
}