# lib/api-zod/src/generated/types/articleListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CasinoKing Blog API
 * OpenAPI spec version: 0.1.0
 */
import type { Article } from "./article";

export interface ArticleListResponse {
  articles: Article[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}