# lib/db/src/schema/articles.ts
import { pgTable, text, boolean, date, uuid, timestamp, numeric } from "drizzle-orm/pg-core";
import { createInsertSchema } from "drizzle-zod";
import { z } from "zod/v4";

export const articlesTable = pgTable("articles", {
  id: uuid("id").primaryKey().defaultRandom(),
  title: text("title").notNull(),
  slug: text("slug").unique().notNull(),
  excerpt: text("excerpt"),
  content: text("content"),
  category: text("category"),
  image: text("image"),
  author: text("author"),
  authorBio: text("author_bio"),
  authorAvatar: text("author_avatar"),
  focusKeyword: text("focus_keyword"),
  metaDescription: text("meta_description"),
  date: date("date"),
  readTime: text("read_time"),
  featured: boolean("featured").default(false),
  published: boolean("published").default(false),
  createdAt: timestamp("created_at", { withTimezone: true }).defaultNow(),
});

export const insertArticleSchema = createInsertSchema(articlesTable).omit({ id: true, createdAt: true });
export type InsertArticle = z.infer<typeof insertArticleSchema>;
export type Article = typeof articlesTable.$inferSelect;